requirements.txt
gradio>=6.0.0
openai>=1.0.0
python-dotenv
beautifulsoup4
requests
